dissolve plotting
add functions to diffraction
